test: move tests from commands to tools
